+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # orchid-core
 
+## 0.1.14
+
+### Patch Changes
+
+- Use gen_random_uuid as a default uuid default
+
 ## 0.1.13
 
 ### Patch Changes