+++ packages/schema-to-zod/CHANGELOG.md
@@ -1,5 +1,13 @@
 # orchid-orm-schema-to-zod
 
+## 0.3.30
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.1.14
+  - pqb@0.10.29
+
 ## 0.3.29
 
 ### Patch Changes